data-lake: Segregate type definitions to a dedicated file
